requirements.txt
engineering-notation==0.10.0
pillow==12.0.0
stringify==0.1.1
tk==0.1.0
